chore: Move Section logic to utils class
