+++ dhis-2/dhis-services/dhis-service-dxf2/src/main/java/org/hisp/dhis/dxf2/util/SectionUtils.java
@@ -1,0 +1,99 @@
+/*
+ * Copyright (c) 2004-2021, University of Oslo
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ * Redistributions of source code must retain the above copyright notice, this
+ * list of conditions and the following disclaimer.
+ *
+ * Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation
+ * and/or other materials provided with the distribution.
+ * Neither the name of the HISP project nor the names of its contributors may
+ * be used to endorse or promote products derived from this software without
+ * specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
+ * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package org.hisp.dhis.dxf2.util;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+
+import org.hisp.dhis.dataelement.DataElement;
+import org.hisp.dhis.dataset.Section;
+import org.springframework.stereotype.Component;
+
+/**
+ * @author Abyot Asalefew Gizaw <abyota@gmail.com>
+ *
+ */
+@Component
+public class SectionUtils
+{
+    public Map<String, Collection<DataElement>> getOrderedDataElementsMap( Section section )
+    {
+        Map<String, Collection<DataElement>> orderedDataElementsMap = new LinkedHashMap<>();
+
+        if ( section == null )
+        {
+            return orderedDataElementsMap;
+        }
+
+        Long currentCategoryCombo = null;
+
+        List<DataElement> groups = new ArrayList<>();
+
+        int counter = 0;
+
+        for ( DataElement dataElement : section.getDataElements() )
+        {
+            Long cc = dataElement.getCategoryCombo().getId();
+
+            if ( currentCategoryCombo == null )
+            {
+                currentCategoryCombo = cc;
+            }
+
+            if ( cc.equals( currentCategoryCombo ) )
+            {
+                groups.add( dataElement );
+            }
+            else
+            {
+                orderedDataElementsMap.put( section.getId() + "-" + currentCategoryCombo + "-" + counter,
+                    new ArrayList<>( groups ) );
+
+                currentCategoryCombo = cc;
+
+                groups = new ArrayList<>();
+
+                groups.add( dataElement );
+
+                counter++;
+            }
+        }
+
+        if ( !orderedDataElementsMap.keySet()
+            .contains( section.getId() + "-" + currentCategoryCombo + "-" + counter ) )
+        {
+            orderedDataElementsMap.put( section.getId() + "-" + currentCategoryCombo + "-" + counter,
+                new ArrayList<>( groups ) );
+        }
+
+        return orderedDataElementsMap;
+    }
+}
